requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
numpy
faiss-cpu
requests
pydantic
google-generativeai
sentence-transformers